+++ complex objects/ios/Toodle/Assets.xcassets/AppIcon.appiconset/Contents.json
