+++ complex objects/ios/Toodle/Rust/Logins/LoginManager.swift
